feat(bot): restructure bot imports and add StructuredBot
- Refactor bot module imports to be more organized and specific to each bot type.
- Rename PydanticBot to StructuredBot to better reflect its functionality.
- Add StructuredBot to the top-level API and update __init__.py accordingly.
- Update import paths in cli/serve.py to reflect the new structure.
- Ensure creation of .llamabot directory at initialization.

BREAKING CHANGE: The import paths for ChatBot, QueryBot, SimpleBot, and ImageBot have changed. Users will need to update their import statements to reflect the new structure.
