+++ llamabot/__init__.py
@@ -10,10 +10,23 @@ Use it to control the top-level API of your Python data science project.
 # Ensure that ~/.llamabotrc exists.
 from pathlib import Path
 
-from .bot import ChatBot, QueryBot, SimpleBot, ImageBot
+from .bot.chatbot import ChatBot
+from .bot.querybot import QueryBot
+from .bot.simplebot import SimpleBot
+from .bot.imagebot import ImageBot
+from .bot.structuredbot import StructuredBot
 from .recorder import PromptRecorder
+from .prompt_manager import prompt
 
-__all__ = ["ChatBot", "ImageBot", "SimpleBot", "QueryBot", "PromptRecorder"]
+__all__ = [
+    "ChatBot",
+    "ImageBot",
+    "SimpleBot",
+    "QueryBot",
+    "PromptRecorder",
+    "StructuredBot",
+    "prompt",
+]
 
 
 (Path.home() / ".llamabot").mkdir(parents=True, exist_ok=True)
